JPA DAO Refactoring

+++ 04-JPA-API/src/br/com/fiap/jpa/dao/impl/CarroDAOImpl.java
@@ -4,6 +4,8 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.jpa.dao.CarroDAO;
 import br.com.fiap.jpa.entity.Carro;
+import br.com.fiap.jpa.exception.CarNotFoundException;
+import br.com.fiap.jpa.exception.CommitException;
 
 public class CarroDAOImpl implements CarroDAO {
 
@@ -17,23 +19,19 @@ public class CarroDAOImpl implements CarroDAO {
 	@Override
 	public void create(Carro carro) {
 		em.persist(carro);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
 	public void update(Carro carro) {
 		em.merge(carro);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
-	public void delete(int codigo) {
+	public void delete(int codigo) throws CarNotFoundException {
 		Carro carro = em.find(Carro.class,codigo);
+		if (carro == null)
+			throw new CarNotFoundException("Registro não encontrado");
 		em.remove(carro);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
@@ -41,6 +39,20 @@ public class CarroDAOImpl implements CarroDAO {
 		return em.find(Carro.class, codigo);
 	}
 
-	
+	@Override
+	public void commit() throws CommitException {
+		try {
+			em.getTransaction().begin();
+			em.getTransaction().commit();
+		}catch(Exception e) {
+			if (em.getTransaction().isActive())
+				em.getTransaction().rollback();
+			e.printStackTrace();
+			throw new CommitException("Erro ao gravar");
+		}
+	}
+
 	
 }
+
+
